% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_4_TER_BASE.R
\name{adj_TER_p}
\alias{adj_TER_p}
\title{adj_TER_p}
\usage{
adj_TER_p(x, pcols, adj_method)
}
\arguments{
\item{x}{Output object of a logit_seq function}

\item{pcols}{A vector specifying the positions of unadjusted p-value columns in x}

\item{adj_method}{P-value adjustment method.
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
"qvalue" calls the \emph{qvalue} package. Other methods are from base R.}
}
\value{
A data frame containing all the original information with the adjusted p-values
appended to the end.
}
\description{
Function to adjust p-values from the TER tests
}
\details{
P-values from differential translational efficiency test on individual genes/transcripts
are correctd for multiple testing.
}
\examples{
fit1 <- Ribolog::logit_seq(rr.v2_dummy[,-1], Ribolog::sample_attributes_dummy, read_type ~ lung_metastasis,
                            as.vector(rr.v2_dummy$transcript), adj_method = 'none', long_output = FALSE)
fit1_qval <- adj_TER_p(fit1, c(2,4), "qvalue")
fit3 <- Ribolog::logit_seq(rr.v2_dummy[,-1], Ribolog::sample_attributes_dummy, read_type ~ lung_metastasis * cell_line_origin,
                            as.vector(rr.v2_dummy$transcript), adj_method = 'none', long_output = FALSE)
fit3_fdr <- adj_TER_p(fit3, c(2,4,6,8), "fdr")
}
