% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{build_empirical_null}
\alias{build_empirical_null}
\title{build_empirical_null}
\usage{
build_empirical_null(
  x,
  design,
  uniqueID,
  groupID,
  outcome = "read_type",
  adj_method = "none"
)
}
\arguments{
\item{x}{A data matrix of RNA and RPF read counts (transcript ID column excluded), as used by \code{\link{logit_seq}}.}

\item{design}{Design matrix of the experiment describing samples and their attributes.}

\item{uniqueID}{A variable (column) of \code{design} defining unique experimental preparations, from each
of which one RNA sample and one RPF sample was derived (usually a replicate name).}

\item{groupID}{A variable (column) of \code{design} indicating which replicates should be grouped together
as replicates of the same biological sample. All experimental units sharing the same \code{groupID} are
compared to each other to generate the empirical null.}

\item{outcome}{The variable determining whether a vector of read counts is RNA or RPF. Default: "read_type".}

\item{adj_method}{P-value adjustment method passed to the underlying \code{\link{logit_seq}} calls.
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: "none".}
}
\value{
A vector of z statistics constituting the empirical null (same as \code{\link{generate_ENZ}}'s output).
}
\description{
Convenience wrapper that generates the empirical null (EN) distribution of TER test z
statistics in one call, from a combined RNA+RPF count matrix and its design matrix. Combines
\code{\link{partition_to_uniques}} (partitioning the data into unique replicate preparations) and
\code{\link{generate_ENZ}} (pooling z statistics from replicate-vs-replicate "homo" comparisons) -
see their documentation for details on the underlying mechanics.
}
\examples{
rr.v2_dummy.enz <- build_empirical_null(rr.v2_dummy[, -1], Ribolog::sample_attributes_dummy, uniqueID = "replicate_name", groupID = "cell_line")
}
