% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_DOCUMENTATION.R
\docType{data}
\name{cdna-datasets}
\alias{cdna-datasets}
\alias{arabidopsis_cdna}
\alias{fly_cdna}
\alias{human_cdna}
\alias{maize_cdna}
\alias{mouse_cdna}
\alias{rat_cdna}
\alias{worm_cdna}
\alias{yeast_cdna}
\alias{zebrafish_cdna}
\title{Bundled cDNA sequences}
\format{
A named list, one element per transcript (named by transcript ID, matching
\code{<organism>_annotation$transcript}), each a character vector of individual nucleotides
(e.g. \code{c("A", "T", "G", ...)}) making up that transcript's cDNA sequence.

An object of class \code{list} of length 27628.

An object of class \code{list} of length 13931.

An object of class \code{list} of length 19038.

An object of class \code{list} of length 39591.

An object of class \code{list} of length 20562.

An object of class \code{list} of length 22245.

An object of class \code{list} of length 20222.

An object of class \code{list} of length 6600.

An object of class \code{list} of length 30140.
}
\usage{
arabidopsis_cdna

fly_cdna

human_cdna

maize_cdna

mouse_cdna

rat_cdna

worm_cdna

yeast_cdna

zebrafish_cdna
}
\description{
cDNA reference sequences, one per model organism, bundled with the package so
\code{\link{load_annotation_and_cdna}} can retrieve them without a separate download. Only one
transcript per gene (the one with the longest CDS) is included, matching the corresponding
\code{<organism>_annotation} table.
}
\keyword{datasets}
