% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP_CELP.R
\name{codon2transcript}
\alias{codon2transcript}
\title{codon2transcript}
\usage{
codon2transcript(tr_codon_read_count_loess_corrected_list, count_type)
}
\arguments{
\item{tr_codon_read_count_loess_corrected_list}{A list of codon level read counts for all samples and transcripts.
It is the second element of a tr_codon_bias_coeff_corrected_count object produced by \code{\link{CELP_bias}}.
It has the following structure:
list$<sample.name>$<transcript.ID> data.frame:
[1] codon_number [2] codon_type [3] aa_type [4] observed_count [5] bias_coefficient [6] corrected_count.}

\item{count_type}{Options: "observed_count", "corrected_count".}
}
\value{
A data frame where the first column is transcript IDs and the remaining columns contain per transcript read counts for all samples.
}
\description{
Function to sum up codon counts per transcript.
}
\details{
Stalling bias correction is performed at the codon level but differential translational efficiency analysis
is ususally performed at transcript level. This function sums up codon level counts (observed or corrected) for each transcript.
}
\examples{
rpf_observed_sum_LMCN <- codon2transcript(tr_codon_bias_coeff_loess_corrected_count_LMCN$tr_codon_read_count_loess_corrected, "observed_count")
rpf_corrected_sum_LMCN <- codon2transcript(tr_codon_bias_coeff_loess_corrected_count_LMCN$tr_codon_read_count_loess_corrected, "corrected_count")
}
