% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP_CELP.R
\name{codon_barplot}
\alias{codon_barplot}
\title{codon_barplot}
\usage{
codon_barplot(reads_psite_list, annotation, fasta_path, outfile = NULL)
}
\arguments{
\item{reads_psite_list}{Output object of a psite_info_rW function}

\item{annotation}{A dataframe denoting the annotations}

\item{fasta_path}{A path to the fasta file}

\item{outfile}{file name to output visualisations}
}
\value{
Barplots for each sample
}
\description{
Function to visualise the codon usage in every sample
}
