% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{comb_cor_p}
\alias{comb_cor_p}
\title{com_cor_p}
\usage{
comb_cor_p(p, w, cm)
}
\arguments{
\item{p}{Array (nxm matrix) of p-values. n: number of features e.g. genes or transcripts, m: number of tests being combined.}

\item{w}{Array (nxm matrix) of weights. n: number of features e.g. genes or transcripts, m: number of tests being combined.}

\item{cm}{Correlation matrix of z-values or any other test statistics of individual tests that are being combined
(mxm). m: number of samples.}
}
\value{
A matrix with three columns: [1] M (combined test statistic) [2] nu (degree of freedom of chi-sq distribution)
[3] p.com (the combined or meta-analytical p-value)
}
\description{
Function to combine correlated p-values
}
\details{
This is based on the the thoery expounded in a 2003 paper by Kepher Makambi.
}
\references{
Makambi, K. 2003. Weighted inverse chi-square method for correlated significance tests.
Journal of Applied Statistics, 30(2): 225-234.
}
