% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{create_te}
\alias{create_te}
\title{create_te}
\usage{
create_te(x, idcolumns = NULL, rnacolumns, rpfcolumns, allow_zero_rpf = FALSE)
}
\arguments{
\item{x}{Input data frame containing RNA-seq or Ribo-seq data.
Rows are genes/transcripts and columns are samples.
The data.frame may contain additional columns for gene/transcript ID or other metadata.}

\item{idcolumns}{A vector specifying the columns to be excluded from the calculations.
These columns may contain the gene/transcript ID or any metadata that need to be preserved.}

\item{rnacolumns}{A vector specifying the columns containing RNA counts}

\item{rpfcolumns}{A vector specifying the columns containing RPF counts}

\item{allow_zero_rpf}{A boolean allowing the total RPF counts of a transcript to be 0}
}
\value{
A data frame containing the original ID columns and the calculated TE columns.
}
\description{
Function to create a TE (translational efficiency) data frame from a combined RNA+RPF data frame.
}
\details{
Translational efficiency is calculated as RPF/RNA.
The number and order of samples must be the same in RNA and RPF columns.
}
\examples{
te.v2 <- Ribolog::create_te(rr.v2_dummy, idcolumns = 1, rnacolumns = c(2:9), rpfcolumns = c(10:17))
}
