% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{fit_empirical_null}
\alias{fit_empirical_null}
\title{fit_empirical_null}
\usage{
fit_empirical_null(enz_dataframe, distribution)
}
\arguments{
\item{enz_dataframe}{A dataframe of the empirical null distribution generated by generate_ENZ function}

\item{distribution}{A keyword for the distribution to be fitted to the empirical null}
}
\value{
Dataframe of statistical results from the fitting
}
\description{
Function to fit the empirical null distribution to a given standard distribution
}
\examples{
rr.v2_dummy.enz <- Ribolog::build_empirical_null(rr.v2_dummy[, -1], Ribolog::sample_attributes_dummy,
                                            uniqueID = "replicate_name", groupID = "cell_line")
enz_fit <- Ribolog::fit_empirical_null(rr.v2_dummy.enz, 'norm')
}
