% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{frame_psite_rW}
\alias{frame_psite_rW}
\title{frame_psite_rW}
\usage{
frame_psite_rW(
  reads_psite_list,
  sample = NULL,
  transcripts = NULL,
  region = "all",
  length_range = "all",
  plot_title = NULL
)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{sample}{Sample to be plotted.}

\item{transcripts}{Vector of transcript names to be included. Default: NULL (includes all transcripts).}

\item{region}{"5utr", "cds", "3utr" or "all". Default: "all".}

\item{length_range}{Range of read lengths to be included. Default: "all".}

\item{plot_title}{Title of the plot. Default: NULL.}
}
\description{
Function to plot the percentage of psites falling into each of the three reading frames (periodicity)
}
\details{
This function produces bar plots of the percentage of psites falling into each of the three possible reading frames,
separately for 5' UTR, CDS and 3' UTR regions. In a good dataset, psites in CDS should be enriched for frame 1.
}
\examples{
frame_psites <- frame_psite_rW(reads_psite_list_LMCN, "CN34_r2_rpf")
}
