% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_6_ORF_CODON.R
\name{glm_deviance_test_p}
\alias{glm_deviance_test_p}
\title{glm_deviance_test_p}
\usage{
glm_deviance_test_p(x)
}
\arguments{
\item{x}{Output of a glm run}
}
\value{
P-value calculated from the chisq test of deviance between the model and its corresponding null.
}
\description{
Calculate the p-value from a deviance test comparing a model to its corresponding null.
}
