% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{load_annotation_and_cdna}
\alias{load_annotation_and_cdna}
\title{load_annotation_and_cdna}
\usage{
load_annotation_and_cdna(organism)
}
\arguments{
\item{organism}{Specify one of the organisms that you'd like to query}
}
\description{
Function to load pre-existing annotation data table and cDNA fasta file
}
