% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{metaprofile_psite_rW}
\alias{metaprofile_psite_rW}
\title{metaprofile_psite_rW}
\usage{
metaprofile_psite_rW(reads_psite_list, annotation, sample,
  scale_factors = NULL, length_range = "all", transcripts = NULL,
  utr5l = 25, cdsl = 50, utr3l = 25, plot_title = NULL)
}
\arguments{
\item{reads_psite_list}{A list of reads and their psite coordinates produced by \code{\link{psite_info_rW}}}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}

\item{sample}{Sample to be plotted.}

\item{length_range}{Range of read lengths to be included. Default: "all".}

\item{transcripts}{Vector of transcript names to be included. Default: NULL (includes all transcripts).}

\item{utr5l}{Length of the 5'UTR upstream of start codon to be included in the plot. Default: 25.}

\item{cdsl}{Length of the cds downstream of start codon and upstream od stop codon to be included in the plot. Default: 50.}

\item{utr3l}{Length of the 3'UTR downstream of stop codon to be included in the plot. Default: 25.}

\item{plot_title}{Title of the plot. Default: NULL.}
}
\description{
Function to plot read metaprofiles around the start and stop codons for a sample
}
\details{
Read metaprofiles (ribosome occupancy plots) visualize the positional distribution of reads mapping around start and stop codons.
}
\examples{
print_rop(reads_psite_list_LMCN, annotation_human_cDNA, "<file.path>/LMCN_RPF_ribosome_occupancy_profiles_from_annotation.pdf")
}
