% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP.R
\name{normalize_median_of_ratios2}
\alias{normalize_median_of_ratios2}
\title{normalize_median_of_ratios2}
\usage{
normalize_median_of_ratios2(expression.data.frame, data.columns)
}
\arguments{
\item{expression.data.frame}{A data frame containing RNA-seq, ribo-seq or similar data.
Rows are genes/transcripts and columns are samples.
The data frame may contain additional columns for gene/transcript ID or other metadata.}

\item{data.columns}{A vector of numbers specifying the columns to be normalized together.}
}
\value{
A data frame the same size as the input data frame where non-normalized counts in the 'data.columns'
are replaced in situ by normalized counts. Column names are not changed.
}
\description{
Function to normalize a RNA-seq or ribo-seq dataset for library size variation
}
\details{
The original columns of non-normalized counts are replaced in situ with normalized counts.
Compare to the method 'normalize_median_of_ratios'.

Use the 'data.columns' argument to exclude gene/transcript ID and other metadata columns from the
calculations or to normalize RNA and RPF counts separately while keeping them in the same data frame.
}
\examples{
rna.observed.normalized2 <- normalize_median_of_ratios2(rna.observed, c(2:9))
rpf.observed.normalized2 <- normalize_median_of_ratios2(rpf.observed, c(2:9))
## The first column of each data frame (transcript ID) is excluded from normalization.
}
