% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{pairs2correlograms}
\alias{pairs2correlograms}
\title{pairs2correlograms}
\usage{
pairs2correlograms(x, outfile = NULL)
}
\arguments{
\item{x}{A list of the results of TER tests between all pairs of samples in a data set produced by \code{\link{TER_all_pairs}}.}

\item{outfile}{The path and name of the output pdf file containing the correlograms (correlation matrix heat maps). Default: NULL.}
}
\value{
A list containing the correlation matrices of equivalent replicate-by-replicate TER tests in a data set.
}
\description{
Function to calculate and plot correlograms from equivalent pairwise TER tests.
}
\details{
The Ribolog TER test can be performed on single replicates per biological sample. In a replicated experiment
such as (sample A: reps A1 and A2 + sample B: reps B1 and B2), correlation coefficients of regression z scores from equivalent tests
(A1 vs B1, A2 vs B1, A1 vs B2, A2 vs B2) are used to evaluate replicate homogeneity and help determine
the minimum advisable number of replicates to achieve reproducibility.
}
\examples{
# Subset to a handful of transcripts so the pairwise fits run quickly.
rr.v2_dummy.split <- Ribolog::partition_to_uniques(rr.v2_dummy[1:200, -1], Ribolog::sample_attributes_dummy, "replicate_name")
rr.v2_dummy.pairwise <- Ribolog::TER_all_pairs(rr.v2_dummy.split, Ribolog::sample_attributes_dummy,
                                          "read_type", "replicate_name", "cell_line", adj_method = "none")
rr.v2_dummy.correlograms <- pairs2correlograms(rr.v2_dummy.pairwise)
}
