% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{pairs2pi0s}
\alias{pairs2pi0s}
\title{pairs2pi0s}
\usage{
pairs2pi0s(x, outfile = NULL)
}
\arguments{
\item{x}{A list of the results of TER tests between all pairs of samples in a data set produced by \code{\link{TER_all_pairs}}.}

\item{outfile}{The path and name of the output pdf file. Default: NULL.}
}
\value{
A data frame describing uniqueIDs and groupIDs of pairs of compared samples, the \code{pair_type ("Homo" or "Hetero")}
and the estimated \emph{pi0} (proportion of null features aka not differentially translated transcripts).
}
\description{
Function to estimate and plot the proportion of null features from pairwise TER tests.
}
\details{
A histogram of \emph{pi0}s is created colored by pair type. If \code{outfile} is given, the histogram will be saved to the pdf file, too.
\code{"Homo"} pairs are expected to have a higher proportion of null features than \code{"Hetero"} pairs.
}
\examples{
# Subset to a handful of transcripts so the pairwise fits run quickly.
rr.v2_dummy.split <- Ribolog::partition_to_uniques(rr.v2_dummy[1:200, -1], Ribolog::sample_attributes_dummy, "replicate_name")
rr.v2_dummy.pairwise <- Ribolog::TER_all_pairs(rr.v2_dummy.split, Ribolog::sample_attributes_dummy,
                                          "read_type", "replicate_name", "cell_line", adj_method = "none")
pi0df <- pairs2pi0s(rr.v2_dummy.pairwise)
}
