% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{partition_to_uniques}
\alias{partition_to_uniques}
\title{partition_to_uniques}
\usage{
partition_to_uniques(x, design, uniqueID)
}
\arguments{
\item{x}{A data frame or matrix containing RNA+RPF count data where each row is a transcript and each column is RNA or RPF counts of one sample.
This object must contain only count data (and not, for example, a transcript ID column).}

\item{design}{Design matrix of the experiment describing samples and their attributes.
The i-th row in the design matrix describes the i-th column in the input data frame \code{x}.}

\item{uniqueID}{A variable (column) of the design matrix defining unique experimental preparations
from each of which one RNA sample and one RPF sample was derived. It corresponds to the highest resolution
(lowest level) of classification of samples in the data set apart from the RNA/RPF distinction
and is usually equal to replicate name in biological experiments.}
}
\value{
A list where each element is a data frame containing RNA and RPF count of one replicate and its attributes from the design matrix.
}
\description{
Function to convert a RNA+RPF data frame to a sample-by-sample list.
}
\examples{
rr.v2_dummy.split <- partition_to_uniques(rr.v2_dummy[,-1], Ribolog::sample_attributes_dummy, "replicate_name")
# The first column of rr.v2_dummy contained transcript IDs and was thus excluded from input.
}
