% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{plot_mirrors_CELP}
\alias{plot_mirrors_CELP}
\title{plot_mirrors_CELP}
\usage{
plot_mirrors_CELP(x, ylim_low_i, ylim_up_i, xlim_low_i = NULL,
  xlim_up_i = NULL, sample)
}
\arguments{
\item{x}{Data frame containing CELP output for a trancript in one sample}

\item{ylim_low_i}{Lower bound on y axis}

\item{ylim_up_i}{Upper bound on y axis}

\item{xlim_low_i}{Lower bound on x axis}

\item{xlim_up_i}{Upper bound on x axis}

\item{sample}{Name of the sample to be plotted}
}
\description{
Functions to visualize bias coefficient for the specified transcript in one sample.
}
