% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{print_period_region}
\alias{print_period_region}
\title{print_period_region}
\usage{
print_period_region(reads_psite_list, outfile = NULL)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{outfile}{The path and name of the output pdf file}
}
\description{
Function to print out psite periodicity by region plots of all samples to a pdf file
}
\details{
These plots illustrate the 3-base (reading frame-dependent) periodicity of reads, separately for 5'UTR, CDS and 3'UTR regions.
In a good dataset, psites in the CDS should be enriched for frame 1.
}
\examples{
print_period_region(reads_psite_list_LMCN, "<file.path>/LMCN_Periodicity_by_region.pdf")
}
