% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{print_period_region_length}
\alias{print_period_region_length}
\title{print_period_region_length}
\usage{
print_period_region_length(reads_psite_list, outfile = NULL, cl = 95)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{outfile}{The path and name of the output pdf file}

\item{cl}{An integer in [1,100] specifying a confidence level to restrict the plot to a sub-range of read lengths.
Use this argument to avoid printing out uncommon read lengths. Default:95.}
}
\description{
Function to print out periodicity by length and region plots of all samples to a pdf file
}
\details{
These plots illustrate the 3-base (reading-frame-dependent) periodicity of reads, separately for 5'UTR, CDS and 3'UTR and stratified by read length.
They can inform the choice of read length range for further analysis of ribo-seq data. Good-length reads are abundant and show enrichment for frame 1 in the CDS region.
}
\examples{
print_period_region_length(reads_psite_list, "<file.path>/Periodicity_by_length_region2.pdf")
}
