% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{print_read_end_heatmap}
\alias{print_read_end_heatmap}
\title{print_read_end_heatmap}
\usage{
print_read_end_heatmap(reads_list, annotation, outfile, cl = 85,
  utr5l = 25, cdsl = 40, utr3l = 25)
}
\arguments{
\item{reads_list}{A read_list object produced by riboWaltz}

\item{annotation}{An annotation data frame listing the transcripts and lengths of their 5'UTR, CDS and 3'UTR segments}

\item{outfile}{The path and name of the output pdf file}

\item{cl}{Confidence level, a number between 0 and 100, the percentage of all reads for which read end heat map is provided.
Use this argument to avoid printing out uncommon read lengths. Default:85.}

\item{utr5l}{Length of 5'UTR segment to be included upstream of translation start site. Default: 25.}

\item{cdsl}{Length of coding sequence to be included downstream of translation start site and upstream of translation termination site. Default: 40.}

\item{utr3l}{Length of 3'UTR segment to be included downstream of translation termination site. Default: 25.}
}
\description{
Function to print out read end heatmaps of all samples from a riboWlatz reads_list object to a pdf file
}
\details{
Heatmaps of reads (stratified by read length) mapping to positions around the start and stop codon provide a visual
sense regarding a reasonable offset for p-site assignment as well as the 3-base periodicity of RPF reads.
}
\examples{
print_read_end_heatmap(reads_list, annotation.df, "<file.path>/RPF_read_end_heatmaps.pdf")
}
