% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{print_rop}
\alias{print_rop}
\title{print_rop}
\usage{
print_rop(reads_psite_list, annotation, outfile)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by riboWaltz}

\item{annotation}{An annotation data frame listing the transcripts and lengths of their 5'UTR, CDS and 3'UTR segments}

\item{outfile}{The path and name of the output pdf file}
}
\description{
Function to print out ribosome occupancy profiles of all samples from a riboWlatz reads_psite_list object to a pdf file
}
\details{
Ribosome occupancy profiles provide visual sense of the positional distribution of reads mapping around start and stop codons.
}
\examples{
print_rop(reads_psite_list, annotation.df, "<file.path>/RPF_ribosome_occupancy_profiles2.pdf")
}
