% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{psite_info_rW}
\alias{psite_info_rW}
\title{psite_info_rW}
\usage{
psite_info_rW(
  reads_list,
  offset,
  site = NULL,
  fastapath = NULL,
  fasta_genome = TRUE,
  refseq_sep = NULL,
  bsgenome = NULL,
  gtfpath = NULL,
  txdb = NULL,
  dataSource = NA,
  organism = NA,
  granges = FALSE
)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}}

\item{offset}{Offset data table produced by \code{\link{psite_rW}}}

\item{site}{(Optional) One or more of \code{"psite"}, \code{"asite"}, \code{"esite"}: which p/a/e-site
codon sequence(s) to attach to each read. Requires either \code{fastapath} or \code{bsgenome}, plus either
\code{gtfpath} or \code{txdb}, to look up transcript sequences. Default: \code{NULL} (no sequence lookup).}

\item{fastapath}{(Optional) Path to a FASTA file of reference sequences, used to retrieve site codon
sequences when \code{site} is given. Only one of \code{fastapath}/\code{bsgenome} should be given.}

\item{fasta_genome}{Logical; whether \code{fastapath} points to a whole-genome FASTA (\code{TRUE}, requiring
\code{gtfpath}/\code{txdb} to extract transcript sequences) or a transcript-level FASTA (\code{FALSE}).
Default: \code{TRUE}.}

\item{refseq_sep}{(Optional) A string separator; if given, each FASTA sequence name is truncated to
everything before the first occurrence of \code{refseq_sep}. Default: \code{NULL} (no truncation).}

\item{bsgenome}{(Optional) Name of a \code{BSgenome} annotation package to use as the genome sequence
source instead of \code{fastapath}. Only one of \code{fastapath}/\code{bsgenome} should be given.}

\item{gtfpath}{(Optional) Path to a GTF file used to build transcript models (via
\code{txdbmaker::makeTxDbFromGFF}) when extracting site codon sequences from a whole genome.
Only one of \code{gtfpath}/\code{txdb} should be given.}

\item{txdb}{(Optional) Name of an installed \code{TxDb} annotation package to use instead of \code{gtfpath}.
Only one of \code{gtfpath}/\code{txdb} should be given.}

\item{dataSource, organism}{Passed to \code{txdbmaker::makeTxDbFromGFF} when \code{gtfpath} is given
(purely descriptive metadata attached to the resulting \code{TxDb}). Default: \code{NA}.}

\item{granges}{Logical. If \code{TRUE}, return the result as a \code{GenomicRanges::GRangesList} instead
of a plain list of data frames. Default: \code{FALSE}.}
}
\description{
Function to add p-site offset information to a reads_list object
}
\details{
This functions adds to each read four pieces of information regarding its p-site: distance from start
of the transcript, distance from start and end of the coding sequence (CDS), and region (5' UTR, CDS or 3' UTR).
}
\examples{
reads_psite_list_LMCN <- psite_info_rW(reads_list_LMCN, psite_offset_LMCN)
}
