% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{psite_to_codon_count}
\alias{psite_to_codon_count}
\title{psite_to_codon_count}
\usage{
psite_to_codon_count(reads_psite_list, length_range, annotation, fasta_file)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{length_range}{Range of read lengths to be included in calculation of per codon RPF counts.
Longer and shorter reads may be excluded because of high uncertainty in offset assignment, lack of expected periodicity, etc.}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}

\item{fasta_file}{The reference fasta file to which reads were mapped.}
}
\value{
A list of lists with the following structure: list$<sample.name>$<transcript.ID> data.frame: [1] codon_number [2] codon_type [3] aa_type [4] observed_count
}
\description{
Function to count reads per codon for each transcript.
}
\examples{
tr_codon_read_count_LMCN <- psite_to_codon_count(reads_psite_list_LMCN, c(24:32), annotation_human_cDNA, "<file.path>/Human.GRC38.96_cDNA_longest_CDS.txt")
}
