% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_PREPROCESSING.R
\name{read_annotation}
\alias{read_annotation}
\title{read_annotation}
\usage{
read_annotation(annotation_file)
}
\arguments{
\item{annotation_file}{An annotation txt file listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
The annotation table has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.}
}
\description{
Function to create an annotation data table from a txt file.
}
\details{
Several \pkg{Ribolog} functions use the transcript annotation table. \code{\link{read_annotation}} reads the annotation
from a source file and stores it in a data table object which can be subsequently used by other functions.
Transcript names and segment lengths in the annotation must correspond to the reference sequences to which the reads were mapped.
We recommend creating the annotation txt file directly from the fasta file of cDNA sequences using the python script
\code{Biomart_cDNA_fasta_to_rW_annotation_and_reheadered_longest_CDS_cDNA_fasta.py} provided with the \pkg{Ribolog} package.
}
\note{
Fasta files of cDNA sequences (only one transcript per gene with the longest CDS) and annotation tables for human and a number of popular model organisms
(\emph{Saccharomyces cerevisiae} yeast, \emph{Zea mays} maize, \emph{Arabidopsis thaliana} thale cress,
\emph{Drosophila melanogaster} fruit fly, \emph{Caenorhabditis elegans} worm, \emph{Danio rerio} zebra fish,
\emph{Mus musculus} mouse and \emph{Rattus norvegicus} rat) are included with the \pkg{Ribolog} package.
}
\examples{
annotation_human_cDNA <- read_annotation("<file.path>/Human.GRC38.96_annotation.txt")
}
