% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_6_ORF_CODON.R
\name{remove_0_1_region_transcripts}
\alias{remove_0_1_region_transcripts}
\title{remove_0_1_region_transcripts}
\usage{
remove_0_1_region_transcripts(data)
}
\arguments{
\item{data}{Dataset containing site read counts. This dataset must have a long shape, meaning that there should be only one
column containing read counts (and it MUST be named "count").
Thus, each row in \code{data} contains the read count for one site - transcript - sample combination.
Other sample attributes beyond sample ID may be recorded in additional variables in this dataset, or provided separately through a design matrix
and a key variable (e.g. sample ID) connecting the \code{data} and \code{design} matrices.}
}
\value{
A subset of the input dataset where all transcripts are guaranteed to have two or more non-zero read counts for transcript regions.
}
\description{
Function to remove transcripts with fewer than two active transcript sites from dataset
}
\details{
This function counts the number of pA sites with non-zero read counts for each transcripts and removes transripts
with fewer than two active pA sites. This is essential to avoid errors when running the regression models.
}
