% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_6_ORF_CODON.R
\name{remove_low_levels}
\alias{remove_low_levels}
\title{remove_low_levels}
\usage{
remove_low_levels(xd, model)
}
\arguments{
\item{xd}{Dataframe output of normalised and filtered counts for transcript and psite region, merged with the design matrix}

\item{model}{Intended model to be used in the regression}
}
\value{
data frame of the filtered transcript counts for each region
}
\description{
Function to remove transcripts with not enough data to run a regression
}
