% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{row_standardize}
\alias{row_standardize}
\title{row_standardize}
\usage{
row_standardize(x, columns)
}
\arguments{
\item{x}{Input data frame containing RNA-seq or Ribo-seq data.
Rows are genes/transcripts and columns are samples.
The data frame may contain additional columns for gene/transcript ID or other metadata.}

\item{columns}{A vector specifying the columns to be included for row-wise standardization}
}
\value{
A data frame where the specified columns from the input are row-standardized and
the rest is intact.
}
\description{
Function to standardize a selected block of a data frame row-wise.
}
\details{
Row mean is subtracted from each element in the row and the result is divided by row standard deviation.
}
\examples{
te.v2 <- Ribolog::create_te(rr.v2_dummy, idcolumns = 1, rnacolumns = c(2:9), rpfcolumns = c(10:17))
te.v2.stnd <- Ribolog::row_standardize(te.v2, columns = c(2:9))
}
