% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{TER_all_pairs}
\alias{TER_all_pairs}
\title{TER_all_pairs}
\usage{
TER_all_pairs(x, design, outcome = "read_type", uniqueID, groupID, adj_method)
}
\arguments{
\item{x}{A sample-by-sample list of RNA and RPF count data and sample attributes produced by \code{\link{partition_to_uniques}}.}

\item{design}{Design matrix of the experiment describing samples and their attributes.}

\item{outcome}{The variable determining whether a vector of read counts is RNA or RPF.
This is usually the name of the response variable in the TER test logistic regression performed through \code{\link{logit_seq}}. Default: \code{"read_type"}.}

\item{uniqueID}{A variable (column) of the design matrix defining unique experimental preparations
from each of which one RNA sample and one RPF sample was derived. It corresponds to the highest resolution
(lowest level) of classification of samples in the data set apart from the RNA/RPF distinction
and is usually equal to replicate name in biological experiments.}

\item{groupID}{A variable (column) of the design matrix indicating which replicates should be grouped together.
All experimental units having the same \code{groupID} will be considered replicates of the same biological sample
(or members of the same group of samples).}

\item{adj_method}{P-value adjustment method.
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
"qvalue" calls the \emph{qvalue} package. Other methods are from base R.}
}
\value{
A list of lists containig the results of all pairwise TER tests. If there are n samples in the input list, the output list will consist of C(n,2) elements.
Each element of the list is in turn a list with four attributes:
- \code{uniqueID}s of the two samples compared
- \code{groupID}s of the two samples compared
- \code{pair_type} (\code{"homo"} if the two \code{groupID}s are equal and \code{"hetero"} otherwise)
- \code{fit} containing the output of the TER test in a data frame. See \code{\link{logit_seq}} for details.
}
\description{
Function to perform the logit TER test between all pairs of samples in a data set.
}
\examples{
# Subset to a handful of transcripts so the pairwise fits run quickly.
rr.v2_dummy.split <- partition_to_uniques(rr.v2_dummy[1:200, -1], Ribolog::sample_attributes_dummy, "replicate_name")
rr.v2_dummy.pairwise <- TER_all_pairs(rr.v2_dummy.split, Ribolog::sample_attributes_dummy, "read_type", "replicate_name",
                                 "cell_line", adj_method = "none")
}
