% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_6_ORF_CODON.R
\name{tr_region_logit_dev}
\alias{tr_region_logit_dev}
\title{tr_region_logit_dev}
\usage{
tr_region_logit_dev(
  data,
  model,
  design = NULL,
  sample_ID = NULL,
  adj_method = "none"
)
}
\arguments{
\item{data}{Dataset containing transcript region read counts. This dataset must have a long shape, meaning that there should be only one
column containing read counts (and it MUST be named "count").
Other sample attributes beyond sample ID may be recorded in additional variables in this dataset, or provided separately through a design matrix
and a key variable (e.g. sample ID) connecting the \code{data} and \code{design} matrices.}

\item{model}{Regression model describing the dependence of region counts on sample attribute(s).}

\item{design}{(optional) Design matrix. A matrix describing sample attributes which can be used as predictors in the regression model.}

\item{sample_ID}{(optional) A key variable connecting the counts dataset (\code{data}) and the design matrix.}

\item{adj_method}{P-value adjustment method. Default: "none".
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
"qvalue" calls the \emph{qvalue} package. Other methods are from base R.}
}
\value{
Deviance test p-values (one per transcript).
}
\description{
Function to evaluate the overall effect of predictors on transcript region count through a deviance test.
}
\examples{
fit.o <- tr_region_logit_dev(tr_regions_df_pivot_norm,
                              psite_region ~ lung_metastasis,
                              sample_attributes,
                              "sample_name")
}
