% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP_CELP.R
\name{visualize_CELP}
\alias{visualize_CELP}
\title{visualize_CELP}
\usage{
visualize_CELP(
  tr_codon_read_count_loess_corrected_list,
  transcript,
  panel_rows = 1,
  panel_cols = 1,
  from_codon = NULL,
  to_codon = NULL,
  outfile = NULL
)
}
\arguments{
\item{tr_codon_read_count_loess_corrected_list}{A list of codon level read counts for all samples and transcripts.
It is the second element of a tr_codon_bias_coeff_corrected_count object produced by \code{\link{CELP_bias}}.
It has the following structure:
list$<sample.name>$<transcript.ID> data.frame:
[1] codon_number [2] codon_type [3] aa_type [4] observed_count [5] bias_coefficient [8] corrected_count.}

\item{transcript}{Name of the transcript to be plotted}

\item{panel_rows}{Number of rows in the case of plotting multiple samples on one page. Default: 1.}

\item{panel_cols}{Number of columns in the case of plotting multiple samples on one page. Default: 1.}

\item{from_codon}{Plot starts from this codon. Use this and to_codon to zoom in on particular regions of the transcript. Default: NULL (plot starts at the start codon).}

\item{to_codon}{Plot ends with this codon. Use this and from_codon to zoom in on particular regions of the transcript. Default: NULL (plot stops at the stop codon).}

\item{outfile}{Path and name of the output pdf file. If it is not provided, plots will be printed to standard output. Default: NULL.}
}
\description{
Function to visualize translational stalling (CELP bias) overlaid on observed and corrected read counts.
}
\details{
This function plots the CELP bias coefficient as a curve overlaid on barplots of observed read counts upward
(positive y) and corrected read counts downward (in the nominal negative y range). This allows visual inspection of the
prominent bias positions and a comparison of read count heterogeneity along the transcript before and after
CELP bias correction.
}
\examples{
visualize_CELP(tr_codon_bias_coeff_loess_corrected_count_LMCN$tr_codon_read_count_loess_corrected, "ENST00000000233", "<file.path>/ENST00000000233.CELP.bias.plots.pdf")
visualize_CELP(tr_codon_bias_coeff_loess_corrected_count_LMCN$tr_codon_read_count_loess_corrected, "ENST00000000233", panel_rows = 2, panel_cols = 4, "<file.path>/ENST00000000233.CELP.bias.all.in.one.page.plots.pdf")
}
