% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_4_TER_BASE.R
\name{adj_TER_p}
\alias{adj_TER_p}
\title{adj_TER_p}
\usage{
adj_TER_p(x, pcols, adj.method)
}
\arguments{
\item{x}{Output object of a logit_seq function}

\item{adj.method}{P-value adjustment method.
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{p.columns}{A vector specifying the positions of unadjusted p-value columns in x}
}
\value{
A data frame containing all the original information with the adjusted p-values
appended to the end.
}
\description{
Function to adjust p-values from the TER tests
}
\details{
P-values from differential translational efficiency test on individual genes/transcripts
are correctd for multiple testing.
}
\examples{
fit.1.qval <- adj_TER_p(fit.1, c(4,8), "qvalue")
fit.2.fdr <- adj_TER_p(fit.2, c(4,8,12,16), "fdr")
}
