% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP.R
\name{bam2count}
\alias{bam2count}
\title{bam2count}
\usage{
bam2count(bamfolder, annotation)
}
\arguments{
\item{bamfolder}{Path to the folder containing bam files (one bam file per sample is expected)}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}
}
\value{
A data frame where the first column contains transcript IDs and the remaining columns contain read counts in imported samples.
}
\description{
Function to generate a read counts table from bam files
}
\details{
This function is designed for bam files generated by mapping to a transcriptome (not a genome).
For each sample, it counts the number of reads mapping to different chromosomes (the 'seqname' bam field).
Annotation must be provided to ensure a complete transcript list, including those with zero counts.
Low count transcripts can be filtered out later using the \code{\link{min_count_filter}} function.
RNA and RPF bam files can be placed in the same folder and imported together using this function if CELP bias correction on RPF counts is not desired.
If CELP correction is desired, RNA and RPF bams should be placed in separate folders.
RNA bams should be imported using this function; RPF bams should be imported using the \code{\link{bamtolist_rW}} function following the CELP workflow.
}
\examples{
rna_count_LMCN <- bam2count(bamfolder = "./Data/Bam/RNA", annotation = annotation_human_cDNA)
rpf_count_LMCN <- bam2count(bamfolder = "./Data/Bam/RPF", annotation = annotation_human_cDNA)
}
