% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{bamtolist_rW}
\alias{bamtolist_rW}
\title{bamtolist_rW}
\usage{
bamtolist_rW(bamfolder, annotation, transcript_align = TRUE,
  name_samples = NULL, indel_threshold = 5, refseq_sep = NULL,
  granges = FALSE)
}
\arguments{
\item{bamfolder}{Path to the folder containing ribo-seq bam files (one bam file per sample is expected).}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}

\item{transcript_align}{A logical argument indicating whether the reads were mapped to transcripts (TRUE) or geneome + gtf (FALSE). Default: TRUE.}

\item{indel_threshold}{Maximum number of indels allowed per read. Default: 5.}

\item{...}{...}
}
\value{
The output is a reads_list object each element of which is a data frame representing one sample. Each row in the data frame contains mapping
information (transcript name and read start and end coordinates) for one read.
}
\description{
Function to convert bam files and an annotation file to a reads_list object.
}
\details{
This function takes a folder of ribo-seq bam files (one per sample) and an annotation table as input, and
produces a reads_list object.
}
\examples{
reads_list_LMCN <- bamtolist_rW("<folder.path>/RPF_sorted_indexed", annotation_human_cDNA)
}
