% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{codon2transcript}
\alias{codon2transcript}
\title{codon2transcript}
\usage{
codon2transcript(tr_codon_read_count_loess_corrected_list, count.type)
}
\arguments{
\item{tr_codon_read_count_loess_corrected_list}{A list of per codon read counts for all samples and transcripts
produced by the generate_rpf_loess function. It has the following structure:
list$<sample.name>$<transcript.ID> data.frame:
[1] codon_number [2] codon_type [3] aa_type [4] observed_count [5] loess_pred [6] loess_pred_by_nz_median [7] bias_coefficient [8] corrected_loess.}

\item{count.type}{Options: "observed_count", "loess_pred", "corrected_loess"}
}
\value{
A data frame where the first column is transcript IDs and the remaining columns contain per transcript read counts for all samples.
}
\description{
Function to sum up codon counts per transcript
}
\details{
Stalling bias correction is performed at the codon level but differential translational efficiency analysis
is ususally performed at transcript level. This function sums up the per codon counts (observed or corrected) on each transcript and
returns per transcript counts.
}
\examples{
rpf.observed.sum <- codon2transcript(tr_codon_bias_coeff_loess_corrected_count$tr_codon_read_count_loess_corrected, "observed_count")
rpf.corrected.sum <- codon2transcript(tr_codon_bias_coeff_loess_corrected_count$tr_codon_read_count_loess_corrected, "corrected_loess")
}
