% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{create_te}
\alias{create_te}
\title{create_te}
\usage{
create_te(x, id.columns = NULL, rna.columns, rpf.columns)
}
\arguments{
\item{x}{Input data frame containing RNA-seq or Ribo-seq data.
Rows are genes/transcripts and columns are samples.
The data.frame may contain additional columns for gene/transcript ID or other metadata.}

\item{id.columns}{A vector specifying the columns to be excluded from the calculations.
These columns may contain the gene/transcript ID or any metadata that need to be preserved.}

\item{rna.columns}{A vector specifying the columns containing RNA counts}

\item{rpf.columns}{A vector specifying the columns containing RPF counts}
}
\value{
A data frame containing the original ID columns and the calculated TE columns.
}
\description{
Function to create a TE (translational efficiency) data frame from a combined RNA+RPF data frame
}
\details{
Translational efficiency is calculated as RPF/RNA.
There must be an equal number of RNA and RPF columns.
}
\examples{
te.m5 <- create_te(rna.rpf.combined.m5, 1, c(2:9), c(10:17))
}
