% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{frame_psite_length_rW}
\alias{frame_psite_length_rW}
\title{frame_psite_length_rW}
\usage{
frame_psite_length_rW(reads_psite_list, sample = NULL,
  transcripts = NULL, region = "all", cl = 95,
  length_range = "all", plot_title = NULL)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{sample}{Sample to be plotted.}

\item{transcripts}{Vector of transcript names to be included. Default: NULL (includes all transcripts).}

\item{region}{"5utr", "cds", "3utr" or "all". Default: "all".}

\item{cl}{An integer in [1,100] specifying a confidence level to restrict the plot to a sub-range of read lengths.
Use this argument to avoid printing out uncommon read lengths. Default:95.}

\item{length_range}{Range of read lengths to be included. Default: "all".}

\item{plot_title}{Title of the plot. Default: NULL.}
}
\description{
Function to plot the percentage of psites falling into each of the three reading frames (periodicity) stratified by length
}
\examples{
print_period_region_length(reads_psite_list_LMCN, "<file.path>/LMCN_Periodicity_by_length_region.pdf")
}
