% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{generate_ENZ}
\alias{generate_ENZ}
\title{generate_ENZ}
\usage{
generate_ENZ(x, design, outcome = "read_type", uniqueID, groupID)
}
\arguments{
\item{x}{A sample-by-sample list of RNA and RPF count data and sample attributes produced by \code{\link{partition_to_uniques}}.}

\item{design}{Design matrix of the experiment describing samples and their attributes.}

\item{outcome}{The variable determining whether a vector of read counts is RNA or RPF.
This is usually the name of the response variable in the TER test logistic regression performed through \code{\link{logit_seq}}. Default: \code{"read_type"}.}

\item{uniqueID}{A variable (column) of the design matrix defining unique experimental preparations
from each of which one RNA sample and one RPF sample was derived. It corresponds to the highest resolution
(lowest level) of classification of samples in the data set apart from the RNA/RPF distinction
and is usually equal to replicate name in biological experiments.}

\item{groupID}{A variable (column) of the design matrix indicating which replicates should be grouped together.
All experimental units having the same \code{groupID} will be considered replicates of the same biological sample
(or members of the same group of samples).}
}
\value{
A vector of z statistics constituting empirical null.
}
\description{
Function to generate empirical null distribution.
}
\details{
In large scale hypothesis testing e.g. genomic data sets, it may be possible to observe the null distribution, instead
of relying the theoretically assumed distribution (standard normal for regression). Ribolog compares replicates of
each biological sample (items with the same \code{groupID}) and pools the z values from those regressions to produce the
empirical null.
}
\examples{
rr_LMCN.v2.enz <- generate_ENZ(x = rr_LMCN.v2.split, design = sample_attributes_LMCN, outcome = "read_type", uniqueID = "replicate_name", groupID = "cell_line")
}
