% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP.R
\name{gm_mean}
\alias{gm_mean}
\title{gm_means}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numbers (may include zero or negative numbers)}
}
\value{
Geometric mean of the vector with the zeros and negative numbers replaced by 1
}
\description{
Function to calculate geometric mean of positive numbers in a vector
}
\details{
Zeros and negative numbers will be treated as 1:
They will not contribute to the sum of logs but will be included in calculation of vector length.
Compare to function: gm_mean_z.
}
