% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_4_TER_BASE.R
\name{logit_seq}
\alias{logit_seq}
\title{logit_seq}
\usage{
logit_seq(x, design, model, n, feature.list = NULL)
}
\arguments{
\item{x}{Input data frame where each column contains RNA or RPF counts from a single sample.
Rows are genes/transcripts.}

\item{design}{Design matrix of the experiment describing samples and their attributes.
i-th row in the design matrix corresponds to the i-th column in the input data frame.}

\item{model}{Regression equation modeling the odds ratio of RPF/RNA counts against the selected design variables (sample attributes)}

\item{n}{Number of estimated parameters (including intercept)}

\item{feature.list}{(Optional) A vector containing IDs of genes/transcripts.
Must have the same length as the row number of input data frame.}
}
\value{
A matrix containing the output of the regression.
Four values are reported for each predictor in the regression 'model' including the intercept:
regression coefficient (beta), standard deviation of the estimated beta, z-score and Wald-test p-value.
Gene/transcript IDs provided by the 'feature.list' argument is added to the output matrix as row names.
}
\description{
Function to perform the logistic regression test for differential translational efficiency
}
\details{
The response variable is always the read type which should be a factor variable with two levels: "RNA" and "RPF".
Translational efficiency of each sample if formulated as the odds of RPF vs. RNA reads.
The change in translational efficiency between samples or based on unit values of any predictor
(design variable) is given by the exponentiated regression coefficient.
Exponentiated intercept gives the TE for the sample with all the attributes at the reference level.
}
\examples{
fit.1 <- logit_seq(rna.rpf.combined.m5[,-1], sample.attributes, read.type ~ lung.metastasis, 2, as.vector(rna.rpf.combined.m5$transcript))
The n=2 parameters include the intercept and the effect of lung.metastasis.
fit.2 <- logit_seq(rna.rpf.combined.m5[,-1], sample.attributes, read.type ~ lung.metastasis*cell.line.origin, 4, as.vector(rna.rpf.combined.m5$transcript))
The n=4 parameters include the intercept, the main effects of lung.metastasis and cell.line.origin and their interaction.
}
