% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{min_count_filter}
\alias{min_count_filter}
\title{min_count_filter}
\usage{
min_count_filter(x, min.count, columns, method = "all")
}
\arguments{
\item{x}{Input data frame containing RNA-seq or Ribo-seq data.
Rows are genes/transcripts and columns are samples.
The data.frame may contain additional columns for gene/transcript ID or other metadata.}

\item{min.count}{A single number (float), minimum RNA or RPF count required for a gene/transcript to be retained.}

\item{columns}{A vector specifying the columns to be considered for minimum count filtering.}

\item{method}{The method of filtering. Options: "all", "average". Default: "all".}
}
\value{
Filtered data.frame containing all the original columns but only the rows (genes)
that pass the filtering criterion.
}
\description{
Function to filter out genes with counts below a minimum in one or more samples.
}
\details{
If method="all" is chosen, a gene passes the filtering only if all samples specified
by the "columns" argument have values >= "min.count". If method="average" is chosen, a gene
passes the filtering if the average count among the specified "columns" is >= "min.count".

Use the columns argument to exclude gene/transcript ID and other metadata columns from the
calculations or to filter RNA and RPF counts separately while keeping them in the same data.frame.
}
\examples{
rna.rpf.combined.m5 <- min_count_filter(rna.rpf.combined, 5, c(2:9))
}
