% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP.R
\name{normalize_median_of_ratios}
\alias{normalize_median_of_ratios}
\title{normalize_median_of_ratios}
\usage{
normalize_median_of_ratios(expression.data.frame, data.columns)
}
\arguments{
\item{expression.data.frame}{A data frame containing RNA-seq, ribo-seq or similar data.
Rows are genes/transcripts and columns are samples.
The data frame may contain additional columns for gene/transcript ID or other metadata.}

\item{data.columns}{A vector of numbers specifying the columns to be normalized together.}
}
\value{
A data frame the same size as the input data frame where non-normalized counts in the 'data.columns'
are replaced in situ by normalized counts. Column names are not changed.
}
\description{
Function to normalize a RNA-seq or ribo-seq dataset for library size variation
}
\details{
The original columns of non-normalized counts are replaced in situ with normalized counts.
Compare with \code{\link{normalize_median_of_ratios_append}}.

Use the \code{data.columns} argument to exclude gene or transcript ID and other metadata columns from the
calculations, and, to normalize RNA and RPF counts separately while keeping them in the same data frame.
}
\examples{
The data set rna_CELP_rpf_count_LMCN contains 17 columns: [1] transcript [2:9] RNA counts [10:17] RPF counts.
RNA and RPF counts are normalized separately.
rna_CELP_rpf_count_norm1_LMCN <- Ribolog::normalize_median_of_ratios(rna_CELP_rpf_count_LMCN, c(2:9))
rna_CELP_rpf_count_norm2_LMCN <- Ribolog::normalize_median_of_ratios(rna_CELP_rpf_count_norm1_LMCN, c(10:17))
}
