% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_2_PREP.R
\name{normalize_median_of_ratios_append}
\alias{normalize_median_of_ratios_append}
\title{normalize_median_of_ratios_append}
\usage{
normalize_median_of_ratios_append(expression.data.frame, data.columns)
}
\arguments{
\item{expression.data.frame}{A data frame containing RNA-seq, ribo-seq or similar data.
Rows are genes/transcripts and columns are samples.
The data frame may contain additional columns for gene/transcript ID or other metadata.}

\item{data.columns}{A vector of numbers specifying the columns to be normalized together.}
}
\value{
An augmented data frame with all the original columns retained and normalized columns
appended to the end. New columns are named 'norm.<old_column>'.
}
\description{
Function to normalize a RNA-seq or ribo-seq dataset for library size variation
}
\details{
The original data columns are retained and normalized columns are added.
Compare with \code{\link{normalize_median_of_ratios2}}.

Use the \code{data.columns} argument to exclude gene or transcript ID and other metadata columns from the
calculations, and, to normalize RNA and RPF counts separately while keeping them in the same data frame.
}
\examples{
rna_observed_normalized_LMCN <- normalize_median_of_ratios_append(rna_rpf_count_LMCN, c(2:9))
}
