% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{pca_qc}
\alias{pca_qc}
\title{pca_qc}
\usage{
pca_qc(
  x,
  n,
  outfile = NULL,
  ID = NULL,
  title = "PCA of samples",
  subtitle = NULL,
  legend_title = "ID",
  title_size = 22,
  axis_title_size = 18,
  axis_text_size = 14,
  point_size = 3,
  point_color = "#1F3864",
  label_size = 4,
  legend_position = "top",
  legend_text_size = 14
)
}
\arguments{
\item{x}{A numeric matrix containing RNA, RPF, TE or some other type of data.
Rows are genes/transcripts and columns are samples.
Gene/transcript IDs and sample names may be given as row names and column names,
respectively (but not as additional non-numeric columns or rows).}

\item{n}{The number of principal components to be plotted}

\item{outfile}{The path and name of the output pdf file containing the PCA plots (optional). Default: NULL.}

\item{ID}{A vector containing group IDs to color-code the samples on the PCA plot
(must correspond to the order of samples in \code{x}) (optional). Default: NULL.}

\item{title}{Plot title. Default: "PCA of samples".}

\item{subtitle}{Plot subtitle. Default: \code{NULL}.}

\item{legend_title}{Legend title used for the \code{ID} colour grouping. Default: "ID".}

\item{title_size}{Font size for the (bold) title. Default: 22.}

\item{axis_title_size}{Font size for the (bold) axis titles. Default: 18.}

\item{axis_text_size}{Font size for the (bold) axis tick labels. Default: 14.}

\item{point_size}{Size of the sample points. Default: 3.}

\item{point_color}{Colour of the sample points. Only used when \code{ID} is \code{NULL} (a single,
uncoloured series); when \code{ID} is given, points are coloured by group using ggplot2's default
discrete palette instead. Default: "#1F3864".}

\item{label_size}{Font size for the sample-name labels placed next to each point. Default: 4.}

\item{legend_position}{Where to place the legend (only shown when \code{ID} is given). Default: "top".}

\item{legend_text_size}{Font size for the (bold) legend title/labels. Default: 14.}
}
\description{
Function to produce PCA output for data visualization and QC purposes.
}
\details{
A summary of the PC analysis is printed out to standard output. Percent variance explained,
shown in each axis label, is rounded to two decimal places.
If outfile is not specified, PCA plots will be printed to standard output i.e. the plots panel in Rstudio.
}
\examples{
te.v2 <- Ribolog::create_te(rr.v2_dummy, idcolumns = 1, rnacolumns = c(2:9), rpfcolumns = c(10:17))
te.v2.stnd <- Ribolog::row_standardize(te.v2, columns = c(2:9))
# The first column of input data (transcript ID) had to be removed to create a full-numeric input dataset.
pca_qc(te.v2.stnd[, -1], n = 2, ID = Ribolog::sample_attributes_dummy$cell_line[c(1:8)])
}
