% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_3_QC_PCA.R
\name{pca_qc}
\alias{pca_qc}
\title{pca_qc}
\usage{
pca_qc(x, n, plotsfile)
}
\arguments{
\item{x}{A numeric matrix containing RNA, RPF, TE or some other type of data.
Rows are genes/transcripts and columns are samples.
Gene/transcript IDs and sample names may be given as row names and column names,
respectively (but not as additional non-numeric columns or rows).}

\item{n}{The number of principal components to be plotted}

\item{plotsfile}{The path and name of the output pdf file containing the PCA plots.}
}
\description{
Function to produce PCA output for data visualization and QC purposes
}
\details{
A summary of the PC analysis is printed out to standard output in addition to creating the plots file.
}
\examples{
pca_qc(te.m5.standardized[,-1], 4, "<file.path>/PCA.QC.te.standardized.pdf")
## The first column of input data (transcript ID) had to be removed to create a full-numeric input dataset.
}
