% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{print_period_region}
\alias{print_period_region}
\title{print_period_region}
\usage{
print_period_region(reads_psite_list, outfile)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by riboWaltz}

\item{outfile}{The path and name of the output pdf file}
}
\description{
Function to print out periodicity by region plots of all samples from a riboWlatz reads_psite_list object to a pdf file
}
\details{
These plots illustrate 3-base (reading-frame-dependent) periodicity of reads, separately for 5'UTR, CDS and 3'UTR.
}
\examples{
print_period_region(reads_psite_list, "<file.path>/Periodicity_by_region2.pdf")
}
