% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{print_period_region_length}
\alias{print_period_region_length}
\title{print_period_region_length}
\usage{
print_period_region_length(reads_psite_list, outfile, cl = 90)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by riboWaltz}

\item{outfile}{The path and name of the output pdf file}

\item{cl}{Confidence level, a number between 0 and 100, the percentage of all reads for which periodicity plots are provided.
Use this argument to avoid printing out uncommon read lengths. Default:90.}
}
\description{
Function to print out periodicity by length and region plots of all samples from a riboWlatz reads_psite_list object to a pdf file
}
\details{
These plots illustrate 3-base (reading-frame-dependent) periodicity of reads, separately for 5'UTR, CDS and 3'UTR and stratified by length.
}
\examples{
print_period_region_length(reads_psite_list, "<file.path>/Periodicity_by_length_region2.pdf")
}
