% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{print_read_end_heatmap}
\alias{print_read_end_heatmap}
\title{print_read_end_heatmap}
\usage{
print_read_end_heatmap(reads_list, annotation, outfile, cl = 85,
  utr5l = 50, cdsl = 50, utr3l = 50)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}

\item{outfile}{The path and name of the output pdf file}

\item{cl}{An integer in [1,100] specifying a confidence level to restrict the plot to a sub-range of read lengths.
Use this argument to avoid printing out uncommon read lengths. Default:85.}

\item{utr5l}{Length of 5'UTR segment to be included upstream of translation start site. Default: 50.}

\item{cdsl}{Length of coding sequence to be included downstream of translation start site and upstream of translation termination site. Default: 50.}

\item{utr3l}{Length of 3'UTR segment to be included downstream of translation termination site. Default: 50.}
}
\description{
Function to print out metaheatmaps of reads around start and stop codons for all samples to a pdf file.
}
\details{
Heatmaps of reads (stratified by read length) mapping to positions around the start and stop codon provide a visual
sense regarding a reasonable offset for p-site assignment as well as the 3-base periodicity of RPF reads.
}
\examples{
print_read_end_heatmap(reads_list_LMCN, annotation_human_cDNA, "<file.path>/LMCN_RPF_read_end_heatmaps.pdf")
}
