% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{print_read_ldist}
\alias{print_read_ldist}
\title{print_read_ldist}
\usage{
print_read_ldist(reads_list, outfile, cl = 99)
}
\arguments{
\item{reads_list}{A read_list object produced by riboWaltz}

\item{outfile}{The path and name of the output pdf file}

\item{cl}{Confidence level, a number between 0 and 100, the percentage of all reads for which distribution of length is provided.
Use this argument to avoid printing out extremely uncommon read lengths. Default:99.}
}
\description{
Function to print out read length distributions of all samples from reads_list object to a pdf file
}
\details{
This function saves the read length distribution plots in a pdf file.
}
\examples{
print_read_ldist(reads_list, "<file.path>/RPF_Read_length_distributions2.pdf")
}
