% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{print_read_ldist}
\alias{print_read_ldist}
\title{print_read_ldist}
\usage{
print_read_ldist(reads_list, outfile, cl = 99)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}.}

\item{outfile}{The path and name of the output pdf file.}

\item{cl}{An integer in [1,100] specifying a confidence level to restrict the plot to a sub-range of read lengths.
Use this argument to avoid printing out extremely uncommon read lengths. Default: 99.}
}
\description{
Function to print out read length distributions of all samples in a reads_list object to a pdf file.
}
\details{
This function saves the read length histograms of all samples in a reads_list object to a pdf file.
}
\examples{
print_read_ldist(reads_list_LMCN, "<file.path>/LMCN_RPF_Read_length_distributions.pdf")
}
