% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{print_rop}
\alias{print_rop}
\title{print_rop}
\usage{
print_rop(reads_psite_list, annotation, outfile)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by \code{\link{psite_info_rW}}}

\item{annotation}{Annotation data table produced by \code{\link{read_annotation}} listing transcript names and lengths of their 5'UTR, CDS and 3'UTR segments.
It has five columns: transcript, l_tr, l_utr5, l_cds and l_utr3.
Transcript names and segment lengths must correspond to the reference sequences to which the reads were mapped.}

\item{outfile}{The path and name of the output pdf file}
}
\description{
Function to print out ribosome occupancy profiles of all samples to a pdf file
}
\details{
Ribosome occupancy profiles visualize the positional distribution of reads mapping around start and stop codons.
}
\examples{
print_rop(LMCN_reads_psite_list, annotation_human_cDNA, "<file.path>/LMCN_RPF_ribosome_occupancy_profiles.pdf")
}
