% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{psite_info_rW}
\alias{psite_info_rW}
\title{psite_info_rW}
\usage{
psite_info_rW(reads_list, offset, site = NULL, fastapath = NULL,
  fasta_genome = TRUE, refseq_sep = NULL, bsgenome = NULL,
  gtfpath = NULL, txdb = NULL, dataSource = NA, organism = NA,
  granges = FALSE)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}}

\item{offset}{Offset data table produced by \code{\link{psite_rW}}}

\item{...}{...}
}
\description{
Function to add p-site offset information to a reads_list object
}
\details{
This functions adds to each read four pieces of information regarding its p-site: distance from start
of the transcript, distance from start and end of the coding sequence (CDS), and region (5' UTR, CDS or 3' UTR).
}
\examples{
reads_psite_list_LMCN <- psite_info_rW(reads_list_LMCN, psite_offset_LMCN)
}
