% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{psite_rW}
\alias{psite_rW}
\title{psite_rW}
\usage{
psite_rW(reads_list, flanking = 6, start = TRUE, extremity = "auto",
  plot = FALSE, plot_dir = NULL, plot_format = "png", cl = 99)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}}

\item{flanking}{Minimum number of nucleotides that have to flank the target codon in both directions. Default: 6.}

\item{start}{Logical argument indicating whether to use the start codon as reference. Default: TRUE.
If set to FALSE, the second to last codon is used.}

\item{extremity}{"5end", "3end" or "auto": the read end on which the correction step should be based. Default: "auto".}

\item{plot}{Whether read-length-specific ribosome occupancy plots are produced. Default: FALSE.}

\item{plot_dir}{The directory where the read-length-specific ribosome occupancy plots are saved. This argument is only
considered if \code{plot = TRUE}.}

\item{plot_format}{"png" or "pdf". It is only considered if \code{plot = TRUE}.}
}
\value{
The output is a data table, containing for all samples and read lengths the percentage of reads in the dataset,
the percentage of reads aligning to the start codon, and the distance of p-site from the two read extremities.
}
\description{
Function to calculate the most likely position of p-site for each read length group.
}
\examples{
psite_offset_LMCN <- psite_rW(reads_list_LMCN)
}
