% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP.R
\name{psite_to_codon_count}
\alias{psite_to_codon_count}
\title{psite_to_codon_count}
\usage{
psite_to_codon_count(reads_psite_list, length_range, annotation,
  fasta.file)
}
\arguments{
\item{reads_psite_list}{A reads_psite_list object produced by riboWaltz}

\item{length_range}{Range of read lengths to be included in calculation of per codon RPF counts.
Longer and shorter reads may be excluded because of high uncertainty in offset assignment, lack of expected periodicity, etc.}

\item{annotation}{An annotation data frame listing the transcripts and lengths of their 5'UTR, CDS and 3'UTR segments}

\item{fasta.file}{The reference fasta file to which reads were mapped.}
}
\value{
A list with the following structure: list$<sample.name>$<transcript.ID> data.frame: [1] codon_number [2] codon_type [3] aa_type [4] observed_count
}
\description{
Function to count reads per codon for each transcript from a reads_psite_list object.
}
