% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_1_CELP_BIAS.R
\name{rlength_distr_rW}
\alias{rlength_distr_rW}
\title{rlength_distr_rW}
\usage{
rlength_distr_rW(reads_list, sample, transcripts = NULL, cl = 99)
}
\arguments{
\item{reads_list}{A reads_list object produced by \code{\link{bamtolist_rW}}.}

\item{sample}{Sample name.}

\item{transcripts}{A vector containing the names of transcripts to be included. Default: NULL (all transcripts are included).}

\item{cl}{An integer in [1,100] specifying a confidence level to restrict the plot to a sub-range of read lengths. Default: 99.
Use this argument to avoid printing out extremely uncommon read lengths. Default:99.}
}
\description{
Function to plot ribo-seq read length distribution of a sample.
}
\details{
This function produces a read length histogram for the specified sample.
}
\examples{
rlength_distr_rW(reads_list_LMCN, "CN34_r1_rpf")
}
