% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULE_5_TER_ALT.R
\name{test_ENZ}
\alias{test_ENZ}
\title{test_ENZ}
\usage{
test_ENZ(x, zcols, enz, method = "fit", keep_data = TRUE)
}
\arguments{
\item{x}{Regression fit object produced by \code{\link{logit_seq}} with option \code{long_output = TRUE},
or any data frame or matrix containing a column
or columns of test statistics (z).}

\item{zcols}{Column(s) in \code{x} containing test z statistics.}

\item{enz}{A vector of z values constituting empirical null e.g. produced by \code{\link{generate_ENZ}}.}

\item{method}{Can be "direct" or "fit". "direct": two-sided p-value of a z is calculated as
the percentage of ENZ larger than abs(z) or smaller than -abs(z). "fit": a normal distribution is
fitted to the empirical null and is used to obtain two-sided p-values. Default: \code{"fit"}.}

\item{keep_data}{A logical variable indicating the original data set should be included in the output. If it is set to
FALSE, only the empirical null p is reported back. Default: TRUE.}
}
\value{
Vector(s) of p-values calculated from an empirical null, reported alone or appended to the original data set.
}
\description{
Function to calculate p-values by empirical null hypothesis testing.
}
\details{
Function \code{\link{logit_seq}} outputs regression coefficients and p-values by defaults. Option
\code{long_output = TRUE} adds SD(beta) and z columns to the output matrix.
}
\examples{
fit1_LMCN_ENZ_p <- test_ENZ(x = fit1_LMCN, enz = rr_LMCN.v2.enz, zcols = 7)
}
